+++ mitra_activitypub/src/handlers/update.rs
@@ -196,6 +196,9 @@ pub async fn handle_update(
     };
     match verify_portable_object(&activity["object"]) {
         Ok(_) => (),
+        Err(AuthenticationError::InvalidObjectID(message)) => {
+            return Err(ValidationError(message).into());
+        },
         Err(AuthenticationError::NotPortable) => (),
         Err(_) => {
             return Err(ValidationError("invalid portable object").into());