+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix incorrect handling of outgoing delivery result.
 - Use compatible ID when setting `inReplyTo` value.
+- Fix incorrect error message during verification of `Create` and `Update` activities.
 
 ## [2.23.0] - 2024-06-23